completed all validations for info processing confid file

+++ src/com/dataconfig/ConfigInformationProcessingData.java
@@ -16,24 +16,31 @@ import com.dataobject.Assigns;
  */
 public class ConfigInformationProcessingData {
 	
+	String regexEmail = "^[A-Za-z](.*)([@]{1})(.{1,})(\\.)(.{1,})";
+	
 	private static java.sql.Date convertUtilToSql(java.util.Date uDate) {
         java.sql.Date sDate = new java.sql.Date(uDate.getTime());
         return sDate;
     }
+	
 	
 	public void updateHotel(String[] args) {
 		Hotel hotelData = new Hotel();
 		hotelData.setId(Integer.parseInt(args[1]));
 		hotelData.setPhone(Integer.parseInt(args[2]));
 		hotelData.setName(args[3]);
-		String name= args[3];
 		hotelData.setAddress(args[4]);
 		hotelData.setCity(args[5]);
 		int oldHotelId=Integer.parseInt(args[6]);
 		int dbFlag = Integer.parseInt(args[7]);
-		if(name !=null){
-		InformationProcessingDAO informationProcessingDAO=new InformationProcessingDAO();
-		informationProcessingDAO.updateHotel(hotelData, oldHotelId, dbFlag);
+		int hotelId = Integer.parseInt(args[1]);
+		String phone = args[2];
+		String name = args[3];
+		
+		
+		if(hotelId > 0 && phone.length() == 10 && phone != null && name != null && name.length() <= 50 && oldHotelId > 0){
+			InformationProcessingDAO informationProcessingDAO=new InformationProcessingDAO();
+			informationProcessingDAO.updateHotel(hotelData, oldHotelId, dbFlag);
 		}else{
 			System.out.println("Not valid data");
 		}
@@ -48,9 +55,14 @@ public class ConfigInformationProcessingData {
 	public void showHotel(String args[]){
 		Hotel hotelData = new Hotel();
 		hotelData.setId(Integer.parseInt(args[1]));
+		int hotelId = Integer.parseInt(args[1]);
 		int dbFlag = Integer.parseInt(args[2]);
-		InformationProcessingDAO informationProcessingDAO=new InformationProcessingDAO();
-		informationProcessingDAO.showHotel(hotelData,dbFlag);
+		if(hotelId > 0){
+			InformationProcessingDAO informationProcessingDAO=new InformationProcessingDAO();
+			informationProcessingDAO.showHotel(hotelData,dbFlag);	
+		}else{
+			System.out.println("Not valid data");
+		}
 	}
 	
 	public void addHotel(String args[]){
@@ -61,16 +73,34 @@ public class ConfigInformationProcessingData {
 		hotelData.setAddress(args[4]);
 		hotelData.setCity(args[5]);
 		int dbFlag = Integer.parseInt(args[6]);
-		InformationProcessingDAO informationProcessingDAO=new InformationProcessingDAO();
-		System.out.println("Added hotel: "+informationProcessingDAO.addHotel(hotelData, dbFlag));
+		
+		int hotelId = Integer.parseInt(args[1]);
+		String phone = args[2];
+		String name = args[3];
+		String address = args[4];
+		String city = args[5];
+		
+		if(hotelId > 0 && phone.length() == 10 && phone != null && name != null && name.length() <= 50 && address != null && address.length() <= 50 && city != null && city.length() <= 20){
+			InformationProcessingDAO informationProcessingDAO=new InformationProcessingDAO();
+			System.out.println("Added hotel: "+informationProcessingDAO.addHotel(hotelData, dbFlag));	
+		}else{
+			System.out.println("Not valid data");
+		}		
 	}
 
 	public void deleteHotel(String[] args) {
 		Hotel hotelData = new Hotel();
 		hotelData.setId(Integer.parseInt(args[1]));
+		
 		int dbFlag = Integer.parseInt(args[2]);
-		InformationProcessingDAO informationProcessingDAO=new InformationProcessingDAO();
-		informationProcessingDAO.deleteHotel(hotelData, dbFlag);
+		int hotelId = Integer.parseInt(args[1]);
+		
+		if(hotelId > 0){
+			InformationProcessingDAO informationProcessingDAO=new InformationProcessingDAO();
+			informationProcessingDAO.deleteHotel(hotelData, dbFlag);
+		}else{
+			System.out.println("Not valid data");
+		}
 	}
 
 	public void addRoom(String[] args) {
@@ -80,8 +110,18 @@ public class ConfigInformationProcessingData {
 		room.setMaxOccu(Integer.parseInt(args[3]));
 		room.setNightRate(Integer.parseInt(args[4]));
 		int dbFlag = Integer.parseInt(args[5]);
-		InformationProcessingDAO informationProcessingDAO=new InformationProcessingDAO();
-		System.out.println("Added Room: "+informationProcessingDAO.addRoom(room, dbFlag));
+		
+		int nightRate = Integer.parseInt(args[4]);
+		int maxOccu = Integer.parseInt(args[3]);
+		int roomNo = Integer.parseInt(args[1]);
+		int hotelId = Integer.parseInt(args[2]);
+		
+		if(roomNo > 0 && hotelId > 0 && maxOccu > 0 && nightRate >= 0){
+			InformationProcessingDAO informationProcessingDAO=new InformationProcessingDAO();
+			System.out.println("Added Room: "+informationProcessingDAO.addRoom(room, dbFlag));
+		}else{
+			System.out.println("Not valid data");
+		}
 	}
 
 	public void updateRoom(String[] args) {
@@ -95,9 +135,19 @@ public class ConfigInformationProcessingData {
 		int oldRoomNum=Integer.parseInt(args[6]);
 		int oldHotelId=Integer.parseInt(args[7]);
 		int dbFlag = Integer.parseInt(args[8]);
-		InformationProcessingDAO informationProcessingDAO=new InformationProcessingDAO();
-		informationProcessingDAO.updateRoom(room, oldRoomNum, oldHotelId, dbFlag);
+		int setAvailability = Integer.parseInt(args[5]);
+		int nightRate = Integer.parseInt(args[4]);
+		int maxOccu = Integer.parseInt(args[3]);
+		int roomNo = Integer.parseInt(args[1]);
+		int hotelId = Integer.parseInt(args[2]);
 		
+		if((setAvailability == 0 || setAvailability == 1) && roomNo > 0 && hotelId > 0 && maxOccu > 0 && nightRate >= 0 &&
+				oldRoomNum > 0 && oldHotelId > 0){
+			InformationProcessingDAO informationProcessingDAO=new InformationProcessingDAO();
+			informationProcessingDAO.updateRoom(room, oldRoomNum, oldHotelId, dbFlag);
+		}else{
+			System.out.println("Not valid data");
+		}
 	}
 
 	public void showRooms(String[] args) {
@@ -111,8 +161,16 @@ public class ConfigInformationProcessingData {
 		room.setRoomNum(Integer.parseInt(args[1]));
 		room.setHotelId(Integer.parseInt(args[2]));
 		int dbFlag = Integer.parseInt(args[3]);
-		InformationProcessingDAO informationProcessingDAO=new InformationProcessingDAO();
-		informationProcessingDAO.deleteRoom(room, dbFlag);
+		
+		int roomNo = Integer.parseInt(args[1]);
+		int hotelId = Integer.parseInt(args[2]);
+		
+		if(roomNo > 0 && hotelId > 0){
+			InformationProcessingDAO informationProcessingDAO=new InformationProcessingDAO();
+			informationProcessingDAO.deleteRoom(room, dbFlag);
+		}else {
+			System.out.println("Not valid data");
+		}
 	}
 
 	public void addStaff(String[] args) throws ParseException {
@@ -127,10 +185,24 @@ public class ConfigInformationProcessingData {
 		staff.setDepartment(args[6]);
 		staff.setTitle(args[7]);
 		staff.setAge(Integer.parseInt(args[8]));
-		int dbFlag = Integer.parseInt(args[9]);
-		InformationProcessingDAO informationProcessingDAO=new InformationProcessingDAO();
-		System.out.println("Added hotel: "+informationProcessingDAO.addStaff(staff, dbFlag));
 		
+		int staffId = Integer.parseInt(args[1]);
+		String phone = args[2];
+		String name = args[3];
+		String address = args[4];
+		String department = args[6];
+		String title = args[7];
+		int age = Integer.parseInt(args[8]);
+		
+		int dbFlag = Integer.parseInt(args[9]);
+		
+		if(staffId > 0 && phone.length() == 10 && phone != null && name != null && name.length() <= 50 && address != null && address.length() <= 50 && 
+				department != null && department.length() <= 20 && title != null && title.length() <= 20 && age > 0 && date != null){
+			InformationProcessingDAO informationProcessingDAO=new InformationProcessingDAO();
+			System.out.println("Added hotel: "+informationProcessingDAO.addStaff(staff, dbFlag));
+		}else {
+			System.out.println("Not valid data");
+		}
 	}
 
 	public void updateStaff(String[] args) throws ParseException {
@@ -139,17 +211,33 @@ public class ConfigInformationProcessingData {
 		staff.setPhone(args[2]);
 		staff.setName(args[3]);
 		staff.setAddress(args[4]);
+		
 		String date=args[5];
+		
 		java.util.Date date1= new SimpleDateFormat("yyyy-MM-dd").parse(date);
 		staff.setDob(convertUtilToSql(date1));
 		staff.setDepartment(args[6]);
 		staff.setTitle(args[7]);
 		staff.setAge(Integer.parseInt(args[8]));
 		
+		int staffId = Integer.parseInt(args[1]);
+		String phone = args[2];
+		String name = args[3];
+		String address = args[4];
+		String department = args[6];
+		String title = args[7];
+		int age = Integer.parseInt(args[8]);
+		
 		int oldStaffId=Integer.parseInt(args[9]);
 		int dbFlag = Integer.parseInt(args[10]);
-		InformationProcessingDAO informationProcessingDAO=new InformationProcessingDAO();
-		informationProcessingDAO.updateStaff(staff, oldStaffId, dbFlag);
+		
+		if(staffId > 0 && phone.length() == 10 && phone != null && name != null && name.length() <= 50 && address != null && address.length() <= 50 && 
+				department != null && department.length() <= 20 && title != null && title.length() <= 20 && age > 0 && date != null){
+			InformationProcessingDAO informationProcessingDAO=new InformationProcessingDAO();
+			informationProcessingDAO.updateStaff(staff, oldStaffId, dbFlag);
+		}else {
+			System.out.println("Not valid data");
+		}
 	}
 
 	public void showStaffs(String[] args) {
@@ -161,9 +249,14 @@ public class ConfigInformationProcessingData {
 	public void deleteStaff(String[] args) {
 		Staff staff = new Staff();
 		staff.setStaffId(Integer.parseInt(args[1]));
+		int staffId = Integer.parseInt(args[1]);
 		int dbFlag = Integer.parseInt(args[2]);
-		InformationProcessingDAO informationProcessingDAO=new InformationProcessingDAO();
-		informationProcessingDAO.deleteStaff(staff, dbFlag);
+		if(staffId > 0){
+			InformationProcessingDAO informationProcessingDAO=new InformationProcessingDAO();
+			informationProcessingDAO.deleteStaff(staff, dbFlag);
+		}else {
+			System.out.println("Not valid data");
+		}
 	}
 
 	public void addCustomer(String[] args) throws ParseException {
@@ -172,12 +265,23 @@ public class ConfigInformationProcessingData {
 		customer.setPhone(args[2]);
 		customer.setName(args[3]);
 		customer.setEmail(args[4]);
+		
+		int customerId = Integer.parseInt(args[1]);
+		String phone = args[2];
+		String name = args[3];
+		String email = args[4];
 		String date=args[5];
+		
 		java.util.Date date1= new SimpleDateFormat("yyyy-MM-dd").parse(date);
 		customer.setDob(convertUtilToSql(date1));
 		int dbFlag = Integer.parseInt(args[6]);
-		InformationProcessingDAO informationProcessingDAO=new InformationProcessingDAO();
-		System.out.println("Added hotel: "+informationProcessingDAO.addCustomer(customer, dbFlag));
+		
+		if(customerId > 0 && phone.length() == 10 && phone != null && name != null && name.length() <= 50 && email != null && email.matches(regexEmail) && email.length() <= 50){
+			InformationProcessingDAO informationProcessingDAO=new InformationProcessingDAO();
+			System.out.println("Added hotel: "+informationProcessingDAO.addCustomer(customer, dbFlag));
+		}else {
+			System.out.println("Not valid data");
+		}	
 	}
 
 	public void updateCustomer(String[] args) throws ParseException {
@@ -191,9 +295,18 @@ public class ConfigInformationProcessingData {
 		customer.setDob(convertUtilToSql(date1));
 		
 		int oldCustomerId=Integer.parseInt(args[6]);
+		int customerId = Integer.parseInt(args[1]);
+		String phone = args[2];
+		String name = args[3];
+		String email = args[4];
 		int dbFlag = Integer.parseInt(args[7]);
-		InformationProcessingDAO informationProcessingDAO=new InformationProcessingDAO();
-		informationProcessingDAO.updateCustomer(customer, oldCustomerId, dbFlag);
+		
+		if(customerId > 0 && phone.length() == 10 && phone != null && name != null && name.length() <= 50 && email != null && email.matches(regexEmail) && email.length() <= 50){
+			InformationProcessingDAO informationProcessingDAO=new InformationProcessingDAO();
+			informationProcessingDAO.updateCustomer(customer, oldCustomerId, dbFlag);
+		}else {
+			System.out.println("Not valid data");
+		}
 	}
 
 	public void showCustomers(String[] args) {
@@ -205,25 +318,40 @@ public class ConfigInformationProcessingData {
 	public void deleteCustomer(String[] args) {
 		Customer customer =  new Customer();
 		customer.setCustomerId(Integer.parseInt(args[1]));
+		int customerId = Integer.parseInt(args[1]);
 		int dbFlag = Integer.parseInt(args[2]);
-		InformationProcessingDAO informationProcessingDAO=new InformationProcessingDAO();
-		informationProcessingDAO.deleteCustomer(customer, dbFlag);
+		if(customerId > 0){
+			InformationProcessingDAO informationProcessingDAO=new InformationProcessingDAO();
+			informationProcessingDAO.deleteCustomer(customer, dbFlag);
+		}else {
+			System.out.println("Not valid data");
+		}
 	}
 
 	public void checkRoomAvailability(String[] args) {
 		int roomNo = Integer.parseInt(args[1]);
 		int hotelId = Integer.parseInt(args[2]);
 		int dbFlag = Integer.parseInt(args[3]);
-		InformationProcessingDAO informationProcessingDAO=new InformationProcessingDAO();
-		informationProcessingDAO.checkRoomAvailability(roomNo, hotelId, dbFlag);
+		if(roomNo > 0 && hotelId > 0){
+			InformationProcessingDAO informationProcessingDAO=new InformationProcessingDAO();
+			informationProcessingDAO.checkRoomAvailability(roomNo, hotelId, dbFlag);
+		}
+		else {
+			System.out.println("Not valid data");
+		}
 	}
 
 	public void checkRoomTypeAvailability(String[] args) {
 		String roomType = args[1];
 		int hotelId = Integer.parseInt(args[2]);
 		int dbFlag = Integer.parseInt(args[3]);
-		InformationProcessingDAO informationProcessingDAO=new InformationProcessingDAO();
-		informationProcessingDAO.checkRoomTypeAvailability(roomType, hotelId, dbFlag);
+		if(hotelId > 0){
+			InformationProcessingDAO informationProcessingDAO=new InformationProcessingDAO();
+			informationProcessingDAO.checkRoomTypeAvailability(roomType, hotelId, dbFlag);
+		}
+		else {
+			System.out.println("Not valid data");
+		}
 	}
 
 	public void setRoomAvailability(String[] args) {
@@ -231,8 +359,12 @@ public class ConfigInformationProcessingData {
 		int roomNo = Integer.parseInt(args[2]);
 		int hotelId = Integer.parseInt(args[3]);
 		int dbFlag = Integer.parseInt(args[4]);
-		InformationProcessingDAO informationProcessingDAO=new InformationProcessingDAO();
-		informationProcessingDAO.setRoomAvailability(availability, roomNo, hotelId, dbFlag);
+		if(availability > 0 && roomNo > 0 && hotelId > 0){
+			InformationProcessingDAO informationProcessingDAO=new InformationProcessingDAO();
+			informationProcessingDAO.setRoomAvailability(availability, roomNo, hotelId, dbFlag);
+		}else {
+			System.out.println("Not valid data");
+		}	
 	}
 
 	public void assignRoom(String[] args) {
@@ -245,7 +377,11 @@ public class ConfigInformationProcessingData {
 		int roomNo = Integer.parseInt(args[4]);
 		int hotelId = Integer.parseInt(args[5]);
 		int dbFlag = Integer.parseInt(args[6]);
-		InformationProcessingDAO informationProcessingDAO = new InformationProcessingDAO();
-		informationProcessingDAO.assignRoom(assigns, staffId, customerId, noOfGuests, roomNo, hotelId, dbFlag);
-	}
+		if(staffId > 0 && customerId > 0 && noOfGuests > 0 && roomNo > 0 && hotelId > 0){
+			InformationProcessingDAO informationProcessingDAO = new InformationProcessingDAO();
+			informationProcessingDAO.assignRoom(assigns, staffId, customerId, noOfGuests, roomNo, hotelId, dbFlag);		
+		}else {
+			System.out.println("Not valid data");
+		}
+		}
 }
